Conflicts:
	README.md

+++ ext/cbson/encoding_helpers.c
@@ -1,19 +1,3 @@
-/*
- * Copyright 2009-2010 10gen, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 #include "encoding_helpers.h"
 
 /*
@@ -87,7 +71,7 @@ static unsigned char isLegalUTF8(const unsigned char* source, int length) {
     return 1;
 }
 
-result_t check_string(const unsigned char* string, const int length,
+result_t check_string(const unsigned char* string, const long length,
                       const char check_utf8, const char check_null) {
     int position = 0;
     /* By default we go character by character. Will be different for checking
